Refactor routes out of index.js
