+++ routes/secure.js
@@ -1,0 +1,3 @@
+export function getSecureData(req, res) {
+    res.json({ message: 'This is secure data.' });
+}
